configure: guard around setting _GNU_SOURCE include/snotypes.h: cast DESCR SPEC sizes to int include/stdio_obj.h: put enum last_direction at top level include/syntab.h: const char * include/io_obj.h: io_write take const char * include/lib.h: io_lib_find take const char * lib/date.c: cast int to enum lib/io.c: io_write, ioo_write, io_lib_find, trypath take const char * lib/pml.c: const char * lib/bsd/inet6.c: true => yes lib/snolib/serv.c: const char * lib/win32/ptyio_obj.c: ptyio_write take const char * lib/win32/inetio_obj.c: inetio_write take const char * lib/stdio_obj.c: stdio_write take const char * lib/auxil/bufio_obj.c: bufio_write take const char * lib/auxil/memio_obj.c: memio_write take const char *

+++ include/snotypes.h
@@ -103,7 +103,7 @@ struct descr {
     VFLD(v);				/* value (new: t) data-type/size */
 };
 
-#define DESCR (sizeof(struct descr))
+#define DESCR ((int)sizeof(struct descr))
 
 /*
  * string specifier (qualifier)
@@ -125,7 +125,7 @@ struct spec {				/* (new: qualifier) */
     struct descr d[2];
 #endif /* SPEC_FIELD_NAMES not defined */
 };
-#define SPEC (sizeof(struct spec))
+#define SPEC ((int)sizeof(struct spec))
 
 /* for generated code which deals with function pointers */
 typedef int (*func_t)(ret_t);
